fix: 插件列表/命令/多级命令/Provider温度四项修复
- 禁用插件后仍在列表中显示（ListMaps 扫描磁盘补充未加载插件）
- 修复命令注册查找使用 Load name 而非 manifest name
- 多级命令节点(/system provider 等)添加默认 handler，直接显示概要
- Go 绑定允许 nil handler 注册纯分组节点
- Provider 温度字段使用 FlexFloat32 兼容前端字符串/数字两种 JSON 格式

+++ internal/api/service/service.go
@@ -222,7 +222,7 @@ func (s *Service) AddProvider(ctx context.Context, c *app.RequestContext) {
 		Endpoint:    data.Endpoint,
 		Token:       data.Token,
 		Model:       data.Model,
-		Temperature: data.Temperature,
+		Temperature: float32(data.Temperature),
 		IsActive:    data.IsActive,
 	}
 
@@ -254,7 +254,7 @@ func (s *Service) AddProvider(ctx context.Context, c *app.RequestContext) {
 			Endpoint:    data.Endpoint,
 			Token:       data.Token,
 			Model:       data.Model,
-			Temperature: data.Temperature,
+			Temperature: float32(data.Temperature),
 		}))
 	}
 
@@ -286,7 +286,7 @@ func (s *Service) UpdateProvider(ctx context.Context, c *app.RequestContext) {
 		Endpoint:    data.Endpoint,
 		Token:       data.Token,
 		Model:       data.Model,
-		Temperature: data.Temperature,
+		Temperature: float32(data.Temperature),
 		IsActive:    data.IsActive,
 	}
 
@@ -298,7 +298,7 @@ func (s *Service) UpdateProvider(ctx context.Context, c *app.RequestContext) {
 		Endpoint:    data.Endpoint,
 		Token:       data.Token,
 		Model:       data.Model,
-		Temperature: data.Temperature,
+		Temperature: float32(data.Temperature),
 	}
 
 	// 运行时同步：先移除同类型旧的，再同步新配置
